Extract filter builders to separate class. Reorganize code in JUnitTestClassExecuter.
